+++ maps/m1.cub
@@ -7,36 +7,17 @@ EA ./textures/lion.xpm
 F 75,108,87
 C 173,216,230
 
-111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000E00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000010000000000000000000000000000000000000000000000110000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000011000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000011000000000000001000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000011000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000011000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000011000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000111000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
-111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111
+111111111111111111111111111
+100000000011000000100000111
+101100000111000000000000111
+101100000000000000000100111
+111111111011000001110000111
+100000000011000001110111111
+111101111111110111000000001
+111101111111110111010000001
+110000001100101110000010011
+100000000000000111000001001
+100000000000000000000101001
+1100000111010001111101110N1
+111101111110101101111010001
+111111111111111111111111111
