+++ src/includes/execution.h
@@ -3,8 +3,8 @@
 
 #include "cub3d.h"
 
-#define WIDTH 1280
-#define HEIGHT 720
+#define WIDTH 800
+#define HEIGHT 800
 #define PLAYER_SPEED 3
 #define BLOCK_SIZE 20
 #define PLAYER_SIZE 10
@@ -65,15 +65,30 @@ typedef struct execution
 } t_execution;
 
 
-
+//ft_execution.c
 void	execution(t_src *src);
-void	init_cub3d(t_execution *cub3d, t_src *src);
 void	exit_failure(char *err_msg);
+float	distance(float x, float y);
+bool	touch(float px, float py, t_execution *execution);
+void	clear_image(t_execution *execution);
 
-void	init_player(t_player *player);
+//ft_player.c
+void	init_player(t_player *player, t_execution *execution);
 int		key_release(int keycode, t_player *player);
 int		key_press(int keycode, t_player *player);
 void	player_movement(t_player *player);
 
+//ft_init_cub3d.c
+void ft_init_cub3d(t_execution *cub3d, t_src *src);
+
+//ft_drawing.c
+void	draw_square(int x, int y, t_execution *execution, int size, int color);
+void	draw_map(t_execution *execution);
+int		draw_a_loop(t_execution *execution);
+void	draw_line(t_player *player, t_execution *execution, float x, int i);
+
+
+//mlx_utils.c
+void	my_mlx_pixel_put(int x, int y, t_execution *execution, int color);
 
 #endif
